requirements.txt
# Core framework
streamlit==1.34.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1

# PDF Parsing
PyMuPDF==1.23.23  # (fitz)
fpdf2==2.7.8

# NLP & ML
spacy==3.7.4
textblob==0.17.1
scikit-learn==1.4.2
sentence-transformers==2.6.1
transformers==4.41.1
torch==2.3.0
numpy==1.24.4
pandas==2.2.2

# Similarity & Parsing
regex==2023.12.25

# Resume Preview
Pillow==10.3.0  # for image handling
base64io==1.0.0

# Optional: for better UX
watchdog==4.0.1

# Optional for CSV/PDF export
fpdf==1.7.2  #

#
